Add min frac to balance labels
min_frac: A minimum required fraction for both pos and neg classes in each slab. If one of the classes in a slab is less than or equal to this fraction, the slab will be ignored (scale=0). You can use frac 0 to make sure there is at least one positive example and frac < 0 to never ignore a slab.