Add 'add_centered' function to batch_request node
The new 'add_centered' function is exactly the same as the previous 'add' function, taking a shape and adding a volume request with that shape centered on the current other batch requests. Also added a new 'add' function to take a roi instead of just a shape that lets you fully choose what batch is actually getting added.