Add `min_frac` functionality to `balance_labels`.
This functionality adds an optional argument that defaults to 0 to the balance_labels node initialization. This argument specifies the minimum fraction for which the balance labels will scale. If the fraction of positive or negative examples goes below this fraction, the scale is set to 0 and that slab is ignored.